+++ changelog.md
@@ -1,5 +1,13 @@
 ## Changelog
 
+### 0.6.0
+
+* Make `Q2SurfaceExtension` public
+
+* Rename `WriteAttempt` to `WriteResult`
+
+* Make Err variant of `ValidationResult` implement `core::error::Error` trait
+
 ### 0.5.2
 
 * More README fixes (: